requirements.txt
Django==1.10.5
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
django-allauth==0.30.0
djangorestframework==3.5.3
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.20.1
greenlet==0.4.12
gunicorn==19.6.0
kombu==4.0.2
psycopg2==2.6.2
python-http-client==2.2.1
pytz==2016.10
redis==2.10.5
sendgrid==3.6.3
sendgrid-django==4.0.4
vine==1.1.3
